requirements.txt
paramiko==3.4.0
GitPython==3.1.40
pyyaml==6.0.1
keyring==24.3.0
python-dateutil==2.8.2
